Rename the top level module to RubyLsp
