+++ lib/ruby-lsp.rb
@@ -1,6 +1,4 @@
 # frozen_string_literal: true
 
-require "syntax_tree"
-
-require_relative "./ruby/lsp/cli"
-require_relative "./ruby/lsp/visitor"
+module RubyLsp
+end
