--- lib/ruby/lsp/requests.rb
@@ -1,9 +1,0 @@
-# frozen_string_literal: true
-
-module Ruby
-  module Lsp
-    module Requests
-      autoload :FoldingRanges, "ruby/lsp/requests/folding_ranges"
-    end
-  end
-end
