+++ lib/ruby_lsp/requests.rb
@@ -1,0 +1,7 @@
+# frozen_string_literal: true
+
+module RubyLsp
+  module Requests
+    autoload :FoldingRanges, "ruby_lsp/requests/folding_ranges"
+  end
+end
