+++ test/test_helper.rb
@@ -3,6 +3,9 @@
 $LOAD_PATH.unshift(File.expand_path("../lib", __dir__))
 
 require "ruby-lsp"
+require "syntax_tree"
+require "ruby_lsp/cli"
+require "ruby_lsp/visitor"
 require "minitest/autorun"
 require "minitest/reporters"
 require "tempfile"
