+++ test/visitor_test.rb
@@ -4,20 +4,20 @@ require "test_helper"
 
 class VisitorTest < Minitest::Test
   def test_can_visit_all_nodes
-    visitor = Ruby::Lsp::Visitor.new
+    visitor = RubyLsp::Visitor.new
 
     SyntaxTree.constants.each do |node|
-      assert_respond_to(visitor, "visit_#{Ruby::Lsp::Visitor.class_to_visit_method(node.to_s)}")
+      assert_respond_to(visitor, "visit_#{RubyLsp::Visitor.class_to_visit_method(node.to_s)}")
     end
   end
 
   def test_class_to_visit_method
-    visit_method_name = Ruby::Lsp::Visitor.class_to_visit_method(SyntaxTree::TStringEnd.name)
+    visit_method_name = RubyLsp::Visitor.class_to_visit_method(SyntaxTree::TStringEnd.name)
     assert_equal("t_string_end", visit_method_name)
   end
 
   def test_visit_tree
-    parsed_tree = Ruby::Lsp::Store::ParsedTree.new(<<~RUBY)
+    parsed_tree = RubyLsp::Store::ParsedTree.new(<<~RUBY)
       class Foo
         def foo; end
 
@@ -34,7 +34,7 @@ class VisitorTest < Minitest::Test
     assert_equal(["Foo", "foo", "Bar", "bar", "baz"], visitor.visited_nodes)
   end
 
-  class DummyVisitor < Ruby::Lsp::Visitor
+  class DummyVisitor < RubyLsp::Visitor
     attr_reader :visited_nodes
 
     def initialize
